Changed behaviour of levels parameters
Degree of generalization is now controlled by the min_cells argument, which specifies the minimum number of cells to remain in the most generalized version of the DEM. This is how the original MrVBF algorithm worked. The default value of 1 results in full generalization, i.e. the input DEM is generalized to 100 percent of its original cell size.